refactor: move emitTypeSpec to emitIR
